// containers/Profile/ProfileDetails/types.ts
// export enum IModalType = 'profile-modal' | 'cover-modal' | 'form-exit-modal';
// export type IDateType = 'startDate' | 'endDate';

export enum IModalType {
  PROFILE_RESIZE_MODAL = 'profile-resize-modal',
  COVER_RESIZE_MODAL = 'cover-resize-modal',
  FORM_EXIT_MODAL = 'form-exit-modal',
}

export enum IDateType {
  START_DATE = 'startDate',
  END_DATE = 'endDate',
}